requirements.txt
fastapi
uvicorn
httpx
beautifulsoup4
python-dotenv
supabase
python-telegram-bot
pydantic
